+++ src/lib.rs
@@ -5,6 +5,7 @@
 pub mod analysis_report;
 pub mod analyzer;
 pub mod args;
+pub mod behaviourless_impl_detector;
 pub mod config;
 pub mod definition_analyzer;
 pub mod humble_adapter_detector;